refactor: 주문 목록과 주문 상세 페이지에서 보이는 OrderListItem 구분
