+++ src/store/OrderState.ts
@@ -6,11 +6,11 @@ export const orderState = atom<OrderItem[]>({
   default: [],
 });
 
-export const currentOrderSelector = selector<OrderItem>({
+export const currentOrderSelector = selector<OrderItem | null>({
   key: 'currentOrderSelector',
   get: ({ get }) => {
     const orderItems = get(orderState);
 
-    return orderItems[orderItems.length - 1];
+    return orderItems ? orderItems[orderItems.length - 1] : null;
   },
 });
